+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## Pending changes
 
-
----
+- [#530](https://github.com/bumble-tech/appyx/issues/530) – **Fixed**: Backstack Parallax motion controller bug
 
 ## 2.0.0-alpha01